Bug fixes in programs, addition of a slide deck to explain the key features of each one.

+++ Wednesday-morning/TutorialsCUDA/program_iic.cu
@@ -280,7 +280,7 @@ int main()
     particleXcoord.HostData[i] = 200.0 * (double)rand() / (double)RAND_MAX;
     particleYcoord.HostData[i] = 200.0 * (double)rand() / (double)RAND_MAX;
     particleZcoord.HostData[i] = 200.0 * (double)rand() / (double)RAND_MAX;
-    particleCharge.HostData[i] = 0.5 - rand() / (double)RAND_MAX;
+    particleCharge.HostData[i] =   0.5 - (double)rand() / (double)RAND_MAX;
   }
 
   // Stage critical constants--see cribSheet struct instance cSh above.
